requirements.txt
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.3.0
requests==2.20.1
elasticsearch==6.3.1
drf-util==0.0.12
drf-yasg==1.17.1
flex==6.13.2
swagger-spec-validator==2.4.1
blinker==1.4
Faker==4.1.0
coverage==5.1
django-nose==1.4.6
django-extensions==2.2.9
jsonschema==3.2.0
PyJWT==1.7.1
django-filter==21.1
pytz==2021.3